batman-adv: throughput meter implementation
The throughput meter module is a simple, kernel-space replacement for
throughtput measurements tool like iperf and netper. It is intended to
approximate TCP behaviour.

It is invoked through batctl: the protocol is connection
oriented, with cumulative acknowledgment and a dynamic-size
sliding window.

The test *can* be interrupted by batctl. A receiver side
timeout avoids unlimited waitings for sender packets: after
one second of inactivity, the receiver abort the ongoing
test.

Signed-off-by: Edo Monticelli <montik@autistici.org>
Signed-off-by: Antonio Quartulli <antonio@open-mesh.com>