batman-adv: use another ICMP packet when sending command from userspace
In the current code a real batadv ICMP packet is used to
communicate from userspace to the module which ICMP
operation should be start (the operation to start depends
on the size of the received packet). This worked good so
fare because the same packet was used to perform the only
two available ICMP operations and the very same packet
received from userspace was also sent over the wire to
perform the operation itself.

As soon as we add new and different ICMP operations this
model will not work well anymore. To improve this feature
make the userspace use a proper packet type to tell the
kernel module which operation to start.

The module will then arrange the rest by itself.

Signed-off-by: Antonio Quartulli <antonio@open-mesh.com>